+++ modules/commons/src/main/java/org/apache/ignite/IgniteInterruptedException.java
