+++ modules/commons/src/main/java/org/apache/ignite/internal/IgniteClientDisconnectedCheckedException.java
