+++ modules/commons/src/main/java/org/apache/ignite/lang/IgniteFutureCancelledException.java
